app/pdf_extractor/requirements.txt
pymupdf>=1.24.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
pillow>=10.0.0
numpy>=1.26.0
pytesseract>=0.3.10
